change of type for duration_seconds from int to float

+++ backend/src/core/video_metadata.py
@@ -7,4 +7,4 @@ class VideoMetaData:
     height: int
     fps: float
     frame_count: int
-    duration_seconds: int
+    duration_seconds: float
